Added dt to controller onframe and other stuff.

+++ src/utilities/timer.h
@@ -22,7 +22,7 @@ class Timer
 		///Stops timer, resetting time. The year will become 0 and we'll all live with jesus christ.
 		void stop();
 
-        void setDesiredFPSRate(float rate);
+        void setDesiredFPSRate(int rate);
         void sleepForTheRestOfTheFrame();
 	private:
 		int mDeltaThreshold;
@@ -30,5 +30,5 @@ class Timer
 		high_resolution_clock::time_point mStartTime;
 		high_resolution_clock mClock;
 		bool mStarted;
-        float fps;
+        int fps;
 };
